simply laziness.
and my morning coffee took longer because of thoughts about yesterday evening.
We live in a relatively small town here. There's a lot of students, though,
and very valuable nodes and networks of sub-cultural people. Yesterday,
a documentary about this sub-culture and the conflicting real estate politics
was premiered on the market place.

It was great to see a lot of these old faces again. Again,
for me, means: after turning into a family man and especially after 24 months
of pandemic *don't-do-anything-with-anybody* mindset which, fortunately,
has started to fade out a bit. Not that everybody was sitting at home but
i certainly had no big part in joint activities.

Now, one of the people interviewed in the film has always been controversial.
He'd some good statements, as usual, but a lot of people in town simply call
him an asshole, sexist, drug addict, a show-off, a creep. I found it fun
to meet him again after some time, amongst hundreds of sub-cultural
friends and supporters, watching this film, that was created to a good part
because of the mobilization of this controversial person and which explains,
based on a particular story ('Die Insel' in Jena) how governments
(state or local) can simple override the desires of people regardless
of the people's years-long efforts to articulate their basic need, in the city
council, in the press, on demonstrations, in official meetings.

Well, at least we can screen a documentary about these misfortunes on the
market place, right? Privately payed and organized, of course.

But all that did not lead to my prolonged morning coffee. After the screenig,
another old-school sub-culture activist approached and asked how i could
still hang about with *this person*, the controversial one. Let's call
the old-schooler **O** and the controversial person **C**. After some
discussion we parted, leaving me with the decission to not ever talk
again with **C** or **O** will never talk with me again.

Background: **C** is organizing small regular demonstrations against
"covid government measures", like obligatory vaccinations, which are not
currently applied and probably not needed anyway, but are discussed often
enough. Now this whole covid/government topic is a sad one anyway and
people, even friends or former friends are fighting about it. Needless
to say that in this little scenario i'm describing, **O** is more on the
*there-is-good-reason-to-be-afraid-by-covid*-side and not so much on the
*but-that-can-not-give-the-government-full-control-over-our-lifes*-side.

Well, it must all be seen in context. And there are different contexts.
Always. I'm not out to fight with people but i like to hear different
and controversial oppinions and a like a bunch of different people that
might not like each other.

But now i'm forced to choose a side or what? **O**, the good guys, or
**C**, the scum and nazi-collaborators. Oh, nazi you say? Well, he did
say *fascists* in particular but we did not clearly define the boundary
between fascist and non-fascist. Sure, why be picky? We don't like
*fascist* nor *nazis*. And i'm simply to despise **C** for letting *fascists*
talk on his demonstrations. Oh, we don't have a *fascist* party in germany
any more. **O** is talking about the AFD youth, the "alternative", which
is, of course, a right-wing political movement, but you could also call it
conservative neo-liberalism that paints itself in the color of resistance
and really picks any topic that the other parties are not willing to
discuss. You can also call them tinfoil heads, which i find a bit sad
because whatever interesting *tinfoil-topic* you have will eventually
be perceived as spread by the right-wings. Sure, that's what they do!
The AFD is increasing it's base through *propaganda*. That's what all the
political parties do, the kind of *propaganda* is a bit different, though.

It's simply tiring. We could dispute so much just about the way one perceives
things. I don't want to dispute. Not in earnest at least. I like to learn
why people think about things the way they do. Why they think a mask
protects us from covid or why they think that this is the most stupid lie
by the government. You know, there are *scientifc facts* in between but
most of the time people follow their guts. Now why should managers and
politicians be excempt from that? They also largly do what they *feel* is
right. And that *right* is a clear *wrong* for someone else.

If i don't officially distance myself from **C**, then **O** will officially
distance himself from me. I just wonder how far this goes. Will i loose all
my friends? I don't think so. But there's another separation, another
parallel world in our little town.

We never ever talk to fascists! It's forbidden by the law, ähh.. i mean
by the agreement of our *scene*.

How is *fascist* defined then? Someone who collaborates with
someone who collaborates with the AFD is clearly a *fascist*.

You know what? If we just had a lever this big to discriminate people
that browse smartphone 8 hours a day?