Yesterday i falsly stated that i'd nothing special to say. In truth, i was just busy with something else. But is this repo really going to become my public diary? I don't know.
Regarding the sad story of **O** vs. **C** and me in between:
I did have a talk with **O**, the next day which was much
less radical. The previous night, we drank a lot (and here's another
retrospective truth: 'simply laziness' was in fact 'laziness
after a couple of beer last night') and the his basic notion
remains, which is, in non-radicalized form:

If someone is collaborating with or supporting people who (for example)
strive for a right-wing political career, then s/he should
not be made comfortable within my own comforting spaces.
If that happens, though, uncomfortable decissions need to be made.