8:3o!
kid was away with grandma at the sea,
i watched a couple of concerts and drank a lot of beer,
finally, during the weekend, when she was back, i slept a lot
and so the first two days we arrived quite early ;)

go see a show of "PoiL Ueda" if you can!

https://www.youtube.com/watch?v=cglVyEFzeZY
https://www.youtube.com/watch?v=yEJknCQdZJU
...

And PoiL, Ni and all them in general!