... so actually, we all have the cough.
One thing that helped regaining some energy was the IGORRR concert last night.
I think this fusion between metal and electronic breakbeat is just taking off.
Oh, there is also a lot of opera singing in this tour. One interesting thing to
see was the actual crowd of people. I've counted about 3 to 4,000 people, most
of them probably coming for Igorrr, while the support bands like Amenra or "Freiheit"
would normally attract metal-headz only, you know, those dark long-haired creatures
that only come to death metal concerts and are never to be seen anywhere else.
So, there where a couple of them to be recognized but all in all the crowd was
a nice mixture of fans of multiple genres. And! The Igorrr guy could have played
two hours more of jungle rythms at the end because the excitement was just exploding
through the roof. Well, it wasn't possible of course because the venue has a
clear contract and staff wants to go to sleep at some point.
Very interesting in any case.